src/core/utils/dateToStringWithTimezone.ts
/**
 * 将日期对象转换为带时区信息的字符串
 * @param date - 要转换的日期对象
 * @returns 带时区信息的字符串
 */
export function dateToStringWithTimezone(date: Date) {
  const offset = -date.getTimezoneOffset()
  const sign = offset >= 0 ? '+' : '-'
  const hours = String(Math.floor(Math.abs(offset) / 60)).padStart(2, '0')
  const minutes = String(Math.abs(offset) % 60).padStart(2, '0')

  const formattedDate =
    date.getFullYear() +
    '-' +
    String(date.getMonth() + 1).padStart(2, '0') +
    '-' +
    String(date.getDate()).padStart(2, '0') +
    'T' +
    String(date.getHours()).padStart(2, '0') +
    ':' +
    String(date.getMinutes()).padStart(2, '0') +
    ':' +
    String(date.getSeconds()).padStart(2, '0') +
    sign +
    hours +
    ':' +
    minutes

  return formattedDate
}
